Consolidate test files into /server/test
